refactor: centralize solver contracts
Move solver contract coverage to arco-contracts and remove obsolete duplicate arco-solver source files so selection, capabilities, request/result, and trait contracts have a single seam. Expand arco-solver-types compatibility re-exports to cover the full contract surface.

Resolves #229
