--- crates/arco-solver/src/config.rs
@@ -1,187 +1,0 @@
-//! Solver configuration types.
-
-use std::collections::BTreeMap;
-
-/// Configuration options for solver behavior.
-#[derive(Debug, Clone, Default, serde::Serialize, serde::Deserialize, PartialEq)]
-pub struct SolverConfig {
-    /// Time limit in seconds. `None` means no limit.
-    pub time_limit: Option<f64>,
-    /// Relative MIP gap tolerance. `None` uses solver default.
-    pub mip_gap: Option<f64>,
-    /// Verbosity level. `None` uses solver default.
-    pub verbosity: Option<u32>,
-    /// Enable/disable presolve. `None` uses solver default.
-    pub presolve: Option<bool>,
-    /// Number of threads to use. `None` uses solver default.
-    pub threads: Option<u32>,
-    /// Feasibility tolerance. `None` uses solver default.
-    pub tolerance: Option<f64>,
-    /// Log solver output to console. `None` uses solver default.
-    pub log_to_console: Option<bool>,
-    /// Family-specific passthrough parameters.
-    #[serde(default)]
-    pub parameters: BTreeMap<String, String>,
-}
-
-impl SolverConfig {
-    /// Create a new configuration with all defaults.
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Set the time limit in seconds.
-    pub fn with_time_limit(mut self, seconds: f64) -> Self {
-        self.time_limit = Some(seconds);
-        self
-    }
-
-    /// Set the relative MIP gap tolerance.
-    pub fn with_mip_gap(mut self, gap: f64) -> Self {
-        self.mip_gap = Some(gap);
-        self
-    }
-
-    /// Set the verbosity level.
-    pub fn with_verbosity(mut self, level: u32) -> Self {
-        self.verbosity = Some(level);
-        self
-    }
-
-    /// Enable or disable presolve.
-    pub fn with_presolve(mut self, enabled: bool) -> Self {
-        self.presolve = Some(enabled);
-        self
-    }
-
-    /// Set the number of threads.
-    pub fn with_threads(mut self, count: u32) -> Self {
-        self.threads = Some(count);
-        self
-    }
-
-    /// Set the feasibility tolerance.
-    pub fn with_tolerance(mut self, tol: f64) -> Self {
-        self.tolerance = Some(tol);
-        self
-    }
-
-    /// Enable or disable console logging.
-    pub fn with_log_to_console(mut self, enabled: bool) -> Self {
-        self.log_to_console = Some(enabled);
-        self
-    }
-
-    /// Set a family-specific passthrough parameter.
-    pub fn with_parameter(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
-        self.parameters.insert(key.into(), value.into());
-        self
-    }
-
-    /// Merge another config onto this one.
-    pub fn merged_with(&self, overlay: &Self) -> Self {
-        let mut merged = self.clone();
-        if overlay.time_limit.is_some() {
-            merged.time_limit = overlay.time_limit;
-        }
-        if overlay.mip_gap.is_some() {
-            merged.mip_gap = overlay.mip_gap;
-        }
-        if overlay.verbosity.is_some() {
-            merged.verbosity = overlay.verbosity;
-        }
-        if overlay.presolve.is_some() {
-            merged.presolve = overlay.presolve;
-        }
-        if overlay.threads.is_some() {
-            merged.threads = overlay.threads;
-        }
-        if overlay.tolerance.is_some() {
-            merged.tolerance = overlay.tolerance;
-        }
-        if overlay.log_to_console.is_some() {
-            merged.log_to_console = overlay.log_to_console;
-        }
-        for (key, value) in &overlay.parameters {
-            merged.parameters.insert(key.clone(), value.clone());
-        }
-        merged
-    }
-
-    /// Check if this configuration is completely empty (all defaults).
-    pub fn is_empty(&self) -> bool {
-        self.time_limit.is_none()
-            && self.mip_gap.is_none()
-            && self.verbosity.is_none()
-            && self.presolve.is_none()
-            && self.threads.is_none()
-            && self.tolerance.is_none()
-            && self.log_to_console.is_none()
-            && self.parameters.is_empty()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_config_new_is_empty() {
-        let config = SolverConfig::new();
-        assert!(config.is_empty());
-    }
-
-    #[test]
-    fn test_config_builder_pattern() {
-        let config = SolverConfig::new()
-            .with_time_limit(60.0)
-            .with_mip_gap(0.01)
-            .with_verbosity(1)
-            .with_presolve(true)
-            .with_threads(4)
-            .with_tolerance(1e-6)
-            .with_log_to_console(false)
-            .with_parameter("solver.option", "enabled");
-
-        assert!(!config.is_empty());
-        assert_eq!(config.time_limit, Some(60.0));
-        assert_eq!(config.mip_gap, Some(0.01));
-        assert_eq!(config.verbosity, Some(1));
-        assert_eq!(config.presolve, Some(true));
-        assert_eq!(config.threads, Some(4));
-        assert_eq!(config.tolerance, Some(1e-6));
-        assert_eq!(config.log_to_console, Some(false));
-        assert_eq!(
-            config.parameters.get("solver.option").map(String::as_str),
-            Some("enabled")
-        );
-    }
-
-    #[test]
-    fn test_config_merge() {
-        let base = SolverConfig::new().with_threads(4).with_time_limit(10.0);
-        let overlay = SolverConfig::new()
-            .with_time_limit(20.0)
-            .with_parameter("k", "v");
-
-        let merged = base.merged_with(&overlay);
-        assert_eq!(merged.threads, Some(4));
-        assert_eq!(merged.time_limit, Some(20.0));
-        assert_eq!(merged.parameters.get("k").map(String::as_str), Some("v"));
-    }
-
-    #[test]
-    fn test_config_clone() {
-        let config1 = SolverConfig::new().with_threads(8);
-        let config2 = config1.clone();
-        assert_eq!(config1.threads, config2.threads);
-    }
-
-    #[test]
-    fn test_config_debug() {
-        let config = SolverConfig::new().with_time_limit(10.0);
-        let debug_str = format!("{config:?}");
-        assert!(debug_str.contains("time_limit"));
-        assert!(debug_str.contains("10.0"));
-    }
-}
